Adds simple coordinate transforms
Moving between cartesian and spherical coordinate systems is something
we might need to do a lot. It makes sense to have theese and possibly
other coordinate transforms, e.g. tralsnations and rotations, in a
dedicated module.

This implements transformations between cartesian and spherical
coordinates, as well as simple tests to make sure that it's actually working.