Adds spherical basis generators
Implementations for the singular and regular spherical basis functions.
These are essential base components, and the implementations have
the secondary purpose of stating the conventions in the package.
Tests are included to chech that the code is internally consistent.
Note that the tests do not guarantee that the calculated values are correct.