Creates submodule for generate methods
The functions implemented in e.g. scipy are quite slow to calculate a full set of  spherical harmonics.
To speed this up we can easily implement some simple recursive relations.
Doing so will need some space in the code, and to avoid namespace clutter and unneccessary long files,
the generate methods are split over multiple files.