Sectorial associated legendre poolynomials
Using a recurrence relation for the associated legendre polynomials is
faster than direct methods. Most implementations start from the
sectorial values, so we need a function to generate those.

The output is one of three noormalizations, the traditional scipy
definition, a normalized version to align with spherical harmonics,
and a stable version where the values are always reasonable in magnitude.