Mark generate submodules as internal
The overall api of the package should not be too deep. The generate modules
are not part of the public api, only a select few functions should be
used in end-user code.

This allows us to use the "same" name for functions in the module and
the "private" module name.