Implements rotations of geometrical vectors
Calculations for transforming spatial vectors (e.g. positions) to
rotated coordinate systems.
The transformation matrix is defined as the matrix needed to calculate
the new coordinates for a given physical point in a new basis.
The inverse (transpose) transformation matrix will map the original
bases to the coefficients of the new bases.