Implements rotations of spherical harmonics expansions
Full 3d rotations of spherical harmonics expansion are implemented
using recurrence relations to obtain the rotation coefficients.
This is confirmed to be stable for at least some fields up to order 60,
and with less than 1% relative error up to order 120.

The conventions are chosen so that a rotated expansion evaluated at
rotated positions give the same values as the original expansion
evaluated as the original positions.

This implementation only supports the "natural" indexing scheme.
Other schemes, e.g. "linear", might be added in the future.