Implements full translations
The most efficient way of doing arbitrary translations is to rotate,
translate coaxially, and rotate back. This is now implemented in two
functions which both operate on sets of expansion coefficients.
One of them is a "simple" function, which takes the new expansion
position, domains etc. The other function takes pre-calculated
re-expansion coefficients for the translation and rotations.
The mode is useful if the same translation has to be applied multiple
times to different fields, e.g. multiple scattering problems.