Classes for bases
Implements classes for the various bases. This is using flat arrays
to store the values instead of rectangular arrays with many zeros.
In some cases normalizations or minor processing is applied in the
component getter. This further allows to store smaller data sets
by exploiting symmetries in the getter, or calculating complex values
only when required.

If these "getter calculations" have large calculation time implications
it should not be too difficult to get rid of them.