Mock of rotation coefficient indexing
Implementation of the indexing scheme for the rotation coefficients.
This is done by storing a different subset of the coefficients
compared to what is currently calculated using the recurrence relation.
Either keep using the same implementation as the current one and manually
move the data around and apply the correct symmetries when storing the
data, or have a look at the relation to reduce the input mode `m` for
positive output modes `p`.