Implements full rotation coefficients
This is done using a subclass of colatitude coefficients, which calculates
the phase needed for the azimuthal rotations only when requested.
This will yield a 50% saving in memory, since we only store the real
colatitude rotation coefficients, and another 50% (for a total of 75%)
since the colatitude coefficients have one additional symmetry.

The phases related to the azimuth rotations are pre-calculated and
stored for efficiency during evaluation. Since this saved only once
per angle, it does not consume large amounts of memory.