Implements rotations
Adds a method to the rotation classes which applies the rotation
coefficients to a set of expansion coefficients. It is possible to
use a pre-existing output location, which can be beneficial if the
rotation is applied multiple times. It is not, however, possible to do
the rotation in place at the moment.