Adds reverse indexing options
Sometimes, especially when debugging, it is nice to have an easy way
to figure out the actual indices corresponding to the linear indices
of a component.

This adds two ways to to that. All the custom indexing methods can now
accept a linear index instead of the actual indices, and will return
the corresponding "normal" indices.
The is also a property which can be used to get a list of the actual
indices of the stored components.