Implements evaluation of coaxial translation coefficients
Uses the same recurrence relations as the previous implementation, but
stored all values in flat arrays instead. This saves about 80% of the
memory in storage, and 90% of the memory during calculation of the
coefficients.
It is significantly slower due to the complicated indexing scheme.
It is certainly possible to optimize the loops to avoid calculating
the linear indices as much as is done at the moment.