Improves argument handling for rotations
The handling of the azimuth rotation angles was not really working as
intended. The old way of changing the angles was to set the angle
property of the rotation object, which is a different way from the
rest of the package.
Now they are set with the `evaluate` method (or when initializing),
the same way as everything else.