Fix for index validity check in coaxial translation
To be able to get the correct value from the indexing of coaxial
translation, it does not matter if the request input and output order
is larger than the set input and output orders. It only matters that
the input order cannot be larger than the minimum of the set input
and output orders (and vice versa for output). This is since we always
calculate and store for input order smaller than output order.