Changes shape form radial bases to be calculated instead of stored
It is more robust of the shape is calculated on the fly instead of
stored on initialization. There might be some change if an object is
reevaluated with a radius of a shape which does not match the shape
that the object was created with. This is considered an edge case which
should hopefully not come up in practice.