Implements reshaping
Allows all the objects to be reshaped.
An implementation is added to the shapes module which attempts to
figure out how to reshape an object so that it can broadcast to the
required shape. This is needed in the cases where the top level object
is created using reduced dataseds for the nested objects.
E.g. evaluating spherical bases and 5 radii, 50 azimuth angles, and 12
colatitude angles. The top level object could have shape (5, 12, 40),
while the radii object has shape (5, 1, 1). The top level object can be
reshaped to (5, 3, 4, 10, 4) and the colatitude object to (1, 3, 4, 1, 1).
This new function makes an attempt to solve this.