Adds implementation of LegendrePolynomials
In some cases the Legendre polynomials are needed. Instead of calculating
the associated Legendre polynomials for all modes, a small class for
just the zero mode polynomials is better. Since some parts of the
implementations are very similar, e.g. normalizations, the associated
Legendre polynomial class is now a subclass of the Legendre class.