Change the definition of which way the rotations work
The translations are designed to translate the field, e.g. move a
source expansion to a new position. The alternative is to reexpand, or
"translate the origin". Now, the rotations are also on the form of
rotating the field, not finding the equivalent expansion in a rotated
system.
This is more natural to work with, since it allows for syntax like that
of the test expansions. As long as everything is documented properly,
both version are possible to do using this package.