Implements transforms for coordinates
It is convenient to have easy and consistent implementations for
transforming the coordinates in the same way we can transform expansions.
This will allow us to choose to either transform the expansions or to
transform the coordinates.
Note that they will behave is the opposite way, i.e. transforming the
expansion as well as the coordinates using the same transform will
give back the same values if the transformed expansion is evaluated at
the transformed coordinates.