Removes reshape methods
The reshape methods were only working in a small set of cases, in
particular it did not manage to shape the objet to the target shape
in all cases, but still did not produce an exception. This is probably
worse than not having the method at all.
If there is a strong need for something similar in the future, it's
probably better to implement something closer to np.expand_dims, which
can add additional dimensions but not change the existing ones.