Changes the meaning of the `xyz` property
Previously this would return the "sparse" representations of each
coordinate. While useful for calculations, it is not very cumbersome
to just get each one of them from the individual properties.
It was, however, quite inconvenient to get the "full" meshes without
getting the stacked mesh and then extracting each coordinate.
Now, the combined property will return the full outputs but still as
individual arrays. This is needed e.g. for visualizing data on a mesh.