Adds source order to circular ring
When requesting the expansion of a translated circular ring source, it's
not guaranteed that it makes sense to use the same order for the original
expansion as for the translated one.
This allows an explicit source order to be used together with the position
argument to control the order of the original expansion.