Changes name for order arguments of translations
The old naming of `input_order` and `output_order` was confusing since
a translation doesn't actually have a distinct input and output order.
The new names are `low_order` and `high_order`, which refers to the
two orders that are calculated for the translation coefficients.

This is reflected in the input arguments of the python wrapper by now
accepting a single `orders` argument, which is either a single int or
a pair of ints. They are automatically converted to the high and low
order values.

An attempt has been done at managing the orders in mixed translations
when they are not the same as the translation object was defined with.